+++ src/JK.Common/Extensions/DateTimeExtensions.cs
@@ -13,7 +13,7 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="days">Number of days to add (can be negative).</param>
         /// <returns>The date the given amount of business days from the start date.</returns>
-        public static DateTime AddWorkDays(this DateTime dt, int days) => new DateTimeHelper().AddWorkDays(dt, days);
+        public static DateTime AddWorkDays(this DateTime dt, int days) => DateHelper.AddWorkDays(dt, days);
 
         /// <summary>
         /// Calculates age of an individual.
@@ -21,14 +21,14 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="birthday">Date of birth.</param>
         /// <returns>Age from birth date to given date.</returns>
-        public static int CalculateAge(this DateTime dt, DateTime birthday) => new DateTimeHelper().CalculateAge(dt, birthday);
+        public static int CalculateAge(this DateTime dt, DateTime birthday) => DateHelper.CalculateAge(dt, birthday);
 
         /// <summary>
         /// Gets the abbreviated day name.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Abbreviated day name.</returns>
-        public static string GetAbbreviatedDayName(this DateTime dt) => new DateTimeHelper().GetAbbreviatedDayName(dt);
+        public static string GetAbbreviatedDayName(this DateTime dt) => DateHelper.GetAbbreviatedDayName(dt);
 
         /// <summary>
         /// Gets the abbreviated day name.
@@ -36,14 +36,14 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="currentCulture">Culture to use when getting day name.</param>
         /// <returns>Abbreviated day name.</returns>
-        public static string GetAbbreviatedDayName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetAbbreviatedDayName(dt, currentCulture);
+        public static string GetAbbreviatedDayName(this DateTime dt, CultureInfo currentCulture) => DateHelper.GetAbbreviatedDayName(dt, currentCulture);
 
         /// <summary>
         /// Gets the abbreviated month name.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Abbreviated month name.</returns>
-        public static string GetAbbreviatedMonthName(this DateTime dt) => new DateTimeHelper().GetAbbreviatedMonthName(dt);
+        public static string GetAbbreviatedMonthName(this DateTime dt) => DateHelper.GetAbbreviatedMonthName(dt);
 
         /// <summary>
         /// Gets the abbreviated month name.
@@ -51,14 +51,14 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="currentCulture">Culture to use when getting month name.</param>
         /// <returns>Abbreviated month name.</returns>
-        public static string GetAbbreviatedMonthName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetAbbreviatedMonthName(dt, currentCulture);
+        public static string GetAbbreviatedMonthName(this DateTime dt, CultureInfo currentCulture) => DateHelper.GetAbbreviatedMonthName(dt, currentCulture);
 
         /// <summary>
         /// Gets the full name of a given day.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Full day name.</returns>
-        public static string GetDayName(this DateTime dt) => new DateTimeHelper().GetDayName(dt);
+        public static string GetDayName(this DateTime dt) => DateHelper.GetDayName(dt);
 
         /// <summary>
         /// Gets the full name of a given day.
@@ -66,14 +66,14 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="currentCulture">Culture to use when getting day name.</param>
         /// <returns>Full day name.</returns>
-        public static string GetDayName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetDayName(dt, currentCulture);
+        public static string GetDayName(this DateTime dt, CultureInfo currentCulture) => DateHelper.GetDayName(dt, currentCulture);
 
         /// <summary>
         /// Gets the full name of a month.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Full month name.</returns>
-        public static string GetMonthName(this DateTime dt) => new DateTimeHelper().GetMonthName(dt);
+        public static string GetMonthName(this DateTime dt) => DateHelper.GetMonthName(dt);
 
         /// <summary>
         /// Gets the full name of a month.
@@ -81,7 +81,7 @@ namespace JK.Common.Extensions
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="currentCulture">Culture to use when getting month name.</param>
         /// <returns>Full month name.</returns>
-        public static string GetMonthName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetMonthName(dt, currentCulture);
+        public static string GetMonthName(this DateTime dt, CultureInfo currentCulture) => DateHelper.GetMonthName(dt, currentCulture);
 
         /// <summary>
         /// Determines whether or not a given date is between (inclusive) the given start and end dates.
@@ -90,20 +90,20 @@ namespace JK.Common.Extensions
         /// <param name="start">Start of date range to check</param>
         /// <param name="end">End of date range to check</param>
         /// <returns>True if date falls within range, otherwise false</returns>
-        public static bool IsBetween(this DateTime dt, DateTime start, DateTime end) => new DateTimeHelper().IsBetween(dt, start, end);
+        public static bool IsBetween(this DateTime dt, DateTime start, DateTime end) => DateHelper.IsBetween(dt, start, end);
 
         /// <summary>
         /// Determines if given date is a weekday.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>True if is a weekday, otherwise false.</returns>
-        public static bool IsWeekday(this DateTime dt) => new DateTimeHelper().IsWeekday(dt);
+        public static bool IsWeekday(this DateTime dt) => DateHelper.IsWeekday(dt);
 
         /// <summary>
         /// Determines if given date is a weekday.
         /// </summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>True if is a weekend, otherwise false.</returns>
-        public static bool IsWeekend(this DateTime dt) => new DateTimeHelper().IsWeekend(dt);
+        public static bool IsWeekend(this DateTime dt) => DateHelper.IsWeekend(dt);
     }
 }